added support to take floating number as input

+++ C++/Area.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int a, b, c, d, y;
+double a, b, c, d, y;
 
 int main()
 {
